+++ src/drive/web/modules/services/components/helpers.js
@@ -2,12 +2,25 @@ import { models } from 'cozy-client'
 
 export const TYPE_DIRECTORY = 'directory'
 
-export const makeNormalizedFile = async (client, folders, file, getIconUrl) => {
+/**
+ * Normalize file for Front usage in <AutoSuggestion> component inside <SearchBar>
+ *
+ * To reduce API call, the fetching of Note URL has been delayed
+ * inside an onSelect function called only if provided to <SearchBar>
+ * see https://github.com/cozy/cozy-drive/pull/2663#discussion_r938671963
+ *
+ * @param client - cozy client instance
+ * @param folders - all the folders returned by API
+ * @param file - file to normalize
+ * @param getIconUrl - method to get icon url
+ * @returns {{path: (*|string), name, icon, id, url: string, onSelect: (function(): Promise<string>)}} file with
+ */
+export const makeNormalizedFile = (client, folders, file, getIconUrl) => {
   const isDir = file.type === TYPE_DIRECTORY
   const dirId = isDir ? file._id : file.dir_id
   const urlToFolder = `${window.location.origin}/#/folder/${dirId}`
 
-  let path, url
+  let path, url, onSelect
   if (isDir) {
     path = file.path
     url = urlToFolder
@@ -15,7 +28,7 @@ export const makeNormalizedFile = async (client, folders, file, getIconUrl) => {
     const parentDir = folders.find(folder => folder._id === file.dir_id)
     path = parentDir && parentDir.path ? parentDir.path : ''
     if (models.file.isNote(file)) {
-      url = await models.note.fetchURL(client, file)
+      onSelect = () => models.note.fetchURL(client, file)
     } else {
       url = `${urlToFolder}/file/${file._id}`
     }
@@ -26,6 +39,7 @@ export const makeNormalizedFile = async (client, folders, file, getIconUrl) => {
     name: file.name,
     path,
     url,
+    onSelect,
     icon: getIconUrl(file)
   }
 }